requirements.txt
Flask==3.1.2
python-dotenv==1.2.1
requests==2.32.5
entsoe-py==0.7.8
pandas==2.3.3
beautifulsoup4==4.14.3